refactor(editor): extract custom cards/badges/views/sections panels
Mechanical module split, step 3/4: the four YAML-passthrough config
panels (incl. item renderers and add/remove/update handlers) move to
panels/CustomConfigPanels.ts. Moved-line deviations per CLAUDE.md
Codacy pitfalls: 'as Record<string, unknown>' instead of 'any' casts,
'catch (error: unknown)' bindings. _sectionDisplayMeta joins the host
contract (used by the custom-cards target dropdown and the section
order panel).
